Update some third-parties to use FetchContent

+++ Crisp/Crisp/Vulkan/VulkanFormatTraits.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <spdlog/spdlog.h>
+#include <Crisp/Common/Logger.hpp>
 #include <vulkan/vulkan.h>
 
 namespace crisp
@@ -8,11 +8,13 @@ namespace crisp
 // Retrieve number of channels for a given format
 template <VkFormat... formats>
 struct NumChannels;
+
 template <>
 struct NumChannels<>
 {
     static constexpr uint32_t value = 0;
 };
+
 template <>
 struct NumChannels<VK_FORMAT_R32G32B32A32_SFLOAT>
 {
@@ -41,7 +43,7 @@ static constexpr uint32_t getSizeOf(VkFormat format)
     case VK_FORMAT_R32G32_SFLOAT:
         return 2 * sizeof(float);
     default:
-        spdlog::critical("Unknown format specified {}", format);
+        spdlog::critical("Unknown format specified {}", static_cast<uint32_t>(format));
     }
 
     return ~0u;
@@ -50,26 +52,31 @@ static constexpr uint32_t getSizeOf(VkFormat format)
 // Retrieve byte size of a vulkan format
 template <VkFormat... formats>
 struct FormatSizeof;
+
 template <>
 struct FormatSizeof<>
 {
     static constexpr size_t value = 0;
 };
+
 template <>
 struct FormatSizeof<VK_FORMAT_R32G32_SFLOAT>
 {
     static constexpr size_t value = 2 * sizeof(float);
 };
+
 template <>
 struct FormatSizeof<VK_FORMAT_R32G32B32_SFLOAT>
 {
     static constexpr size_t value = 3 * sizeof(float);
 };
+
 template <>
 struct FormatSizeof<VK_FORMAT_R32G32B32A32_SFLOAT>
 {
     static constexpr size_t value = 4 * sizeof(float);
 };
+
 template <VkFormat F, VkFormat... Fs>
 struct FormatSizeof<F, Fs...>
 {
@@ -79,16 +86,19 @@ struct FormatSizeof<F, Fs...>
 // Retrieve total byte size for passed types
 template <typename... Ts>
 struct AggregateSizeof;
+
 template <>
 struct AggregateSizeof<>
 {
     static constexpr size_t value = 0;
 };
+
 template <typename T>
 struct AggregateSizeof<T>
 {
     static constexpr size_t value = sizeof(T);
 };
+
 template <typename T, typename... Ts>
 struct AggregateSizeof<T, Ts...>
 {
